Scholarships Refactor
